refactor nifti masker family
