+++ nilearn/input_data/__init__.py
@@ -4,4 +4,5 @@ tools to preprocess neuro-imaging data.
 """
 from .nifti_masker import NiftiMasker
 from .multi_nifti_masker import MultiNiftiMasker
-from .nifti_region import NiftiLabelsMasker, NiftiMapsMasker
+from .nifti_labels_masker import NiftiLabelsMasker
+from .nifti_maps_masker import NiftiMapsMasker
